Super big update part 517
So the Y vector is gone... thats a good thing bc fmincon needs to play with it apparently. Everything is normalised now. Everything is cleaner now. Say byebye to Yst, chord, and their friends, now everything is in the new lift and moment distribution coefficients. this helps to go from like 90 variables to 44!  Q3D still a bit whack tho with the NaNs.